YM2151 updates: - add YM2151Synth::defaultChannelVolumeUpdate() and YM2151Synth::defaultChannelPanUpdate() - add YM2151MidiChannelState::reset() - add updating RL and FL in changePreset()

+++ src/synths/ym2151/OPMFileLoader.h
@@ -36,7 +36,7 @@ struct OPMCPSParams {
 
 struct OPMPatch {
   std::string name;
-  uint8_t number;
+  uint16_t number;
   OPMLFOParams lfoParams;
   OPMChannelParams channelParams;
   OPMOpParams opParams[4];
